Add service number field to apartments and flats
This update introduces a new `service_number` column in the apartments and flats tables for utility service tracking. The UI has been updated to display and edit the service number in relevant forms and detail pages. Additionally, a test page has been created to facilitate testing of the new service number functionality.